style : 구두 리뷰 사항 반영 & organization 도메인 -> admin 도메인으로 이름 변경
- id와 연관관계 어노테이션 기본 생성자로 이동
- id 컬럼 이름 지정 제거
- enum을 string으로 저장
- generatedValue를 identity로 지정
- id 기본값 설정
- 도메인 별 exception 제거
