+++ src/main/kotlin/uket/domain/admin/entity/Organization.kt
@@ -1,16 +1,14 @@
-package uket.uket.domain.terms
+package uket.uket.domain.organization.entity
 
 import jakarta.persistence.*
 import uket.uket.domain.BaseTimeEntity
 
 @Entity
-@Table(name = "document")
-class Document(
+@Table(name = "organization")
+class Organization(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     val id: Long = 0L,
-    val documentNo: Long,
     val name: String,
-    val link: String,
-    val version: Long,
+    val organizationImagePath: String?,
 ) : BaseTimeEntity()
