--- src/main/kotlin/uket/domain/organization/entity/Organization.kt
@@ -1,21 +1,0 @@
-package uket.uket.domain.organization.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.Id
-import jakarta.persistence.Table
-import uket.uket.domain.BaseTimeEntity
-
-@Entity
-@Table(name = "organization")
-class Organization(
-    _id: Long,
-    val name: String,
-    val organizationImagePath: String?,
-) : BaseTimeEntity() {
-    @Id
-    @GeneratedValue
-    @Column(name = "organization_id")
-    val id: Long = _id
-}
