--- src/main/kotlin/uket/domain/organization/exception/OrganizationException.kt
@@ -1,8 +1,0 @@
-package uket.uket.domain.organization.exception
-
-import uket.uket.common.BaseException
-import uket.uket.common.ErrorCode
-
-class OrganizationException(
-    errorCode: ErrorCode,
-) : BaseException(errorCode)
