+++ src/main/kotlin/uket/domain/terms/Terms.kt
@@ -1,23 +1,17 @@
 package uket.uket.domain.terms
 
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.Id
-import jakarta.persistence.Table
+import jakarta.persistence.*
 import uket.uket.domain.BaseTimeEntity
 
 @Entity
 @Table(name = "terms")
 class Terms(
-    _id: Long,
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0L,
     val name: String,
+    @Enumerated(EnumType.STRING)
     val termsType: TermsType,
     val documentNo: Long,
     val isActive: Boolean,
-) : BaseTimeEntity() {
-    @Id
-    @GeneratedValue
-    @Column(name = "terms_id")
-    val id: Long = _id
-}
+) : BaseTimeEntity()
