+++ src/main/kotlin/uket/domain/uketevent/entity/UketEvent.kt
@@ -1,10 +1,6 @@
 package uket.uket.domain.uketevent.entity
 
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.Id
-import jakarta.persistence.Table
+import jakarta.persistence.*
 import uket.uket.domain.BaseTimeEntity
 import uket.uket.domain.uketevent.enums.EventType
 import java.time.LocalDateTime
@@ -12,17 +8,15 @@ import java.time.LocalDateTime
 @Entity
 @Table(name = "uket_event")
 class UketEvent(
-    _id: Long,
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0L,
     val organizationId: Long,
     val name: String,
+    @Enumerated(EnumType.STRING)
     val eventType: EventType,
     val location: String,
     val eventImagePath: String?,
     val displayEndDate: LocalDateTime,
     val ticketPrice: Int,
-) : BaseTimeEntity() {
-    @Id
-    @GeneratedValue
-    @Column(name = "uket_event_id")
-    val id: Long = _id
-}
+) : BaseTimeEntity()
