+++ src/main/kotlin/uket/domain/uketevent/entity/UketEventRound.kt
@@ -1,30 +1,19 @@
 package uket.uket.domain.uketevent.entity
 
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.GeneratedValue
-import jakarta.persistence.Id
-import jakarta.persistence.JoinColumn
-import jakarta.persistence.ManyToOne
-import jakarta.persistence.Table
+import jakarta.persistence.*
 import uket.uket.domain.BaseTimeEntity
 import java.time.LocalDateTime
 
 @Entity
 @Table(name = "uket_event_round")
 class UketEventRound(
-    _id: Long,
-    _uketEvent: UketEvent,
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    val id: Long = 0L,
+    @ManyToOne
+    @JoinColumn(name = "uket_event_id", nullable = false)
+    val uketEvent: UketEvent,
     val name: String,
     val eventDate: LocalDateTime,
     val ticketingDateTime: LocalDateTime,
-) : BaseTimeEntity() {
-    @Id
-    @GeneratedValue
-    @Column(name = "uket_event_round_id")
-    val id: Long = _id
-
-    @ManyToOne
-    @JoinColumn(name = "uket_event_id", nullable = false)
-    val uketEvent: UketEvent = _uketEvent
-}
+) : BaseTimeEntity()
