--- src/main/kotlin/uket/domain/uketevent/exception/UketEventException.kt
@@ -1,8 +1,0 @@
-package uket.uket.domain.uketevent.exception
-
-import uket.uket.common.BaseException
-import uket.uket.common.ErrorCode
-
-class UketEventException(
-    errorCode: ErrorCode,
-) : BaseException(errorCode)
