--- src/main/kotlin/uket/domain/user/exception/UserException.kt
@@ -1,8 +1,0 @@
-package uket.uket.domain.user.exception
-
-import uket.uket.common.BaseException
-import uket.uket.common.ErrorCode
-
-class UserException(
-    errorCode: ErrorCode,
-) : BaseException(errorCode)
